diff_drive: factorize (future) common code in tests
