+++ diff_drive_controller/test/diff_drive_test.cpp
@@ -27,60 +27,7 @@
 
 /// \author Bence Magyar
 
-#include <cmath>
-
-#include <gtest/gtest.h>
-
-#include <ros/ros.h>
-
-#include <geometry_msgs/Twist.h>
-#include <nav_msgs/Odometry.h>
-#include <tf/tf.h>
-
-// Floating-point value comparison threshold
-const double EPS = 0.01;
-const double POSITION_TOLERANCE = 0.01; // 1 cm-s precision
-const double ORIENTATION_TOLERANCE = 0.03; // 0.57 degree precision
-
-class DiffDriveControllerTest : public ::testing::Test
-{
-public:
-
-  DiffDriveControllerTest()
-    : cmd_pub(nh.advertise<geometry_msgs::Twist>("cmd_vel", 100)),
-      odom_sub(nh.subscribe("odom", 100, &DiffDriveControllerTest::odomCallback, this))
-  {
-  }
-
-  ~DiffDriveControllerTest()
-  {
-    odom_sub.shutdown();
-  }
-
-  nav_msgs::Odometry getLastOdom(){ return last_odom; }
-  void publish(geometry_msgs::Twist cmd_vel){ cmd_pub.publish(cmd_vel); }
-  bool isControllerAlive(){ return (odom_sub.getNumPublishers() > 0) && (cmd_pub.getNumSubscribers() > 0); }
-
-private:
-  ros::NodeHandle nh;
-  ros::Publisher cmd_pub;
-  ros::Subscriber odom_sub;
-  nav_msgs::Odometry last_odom;
-
-  void odomCallback(const nav_msgs::Odometry& odom)
-  {
-    ROS_INFO_STREAM("Callback reveived: pos.x: " << odom.pose.pose.position.x
-                     << ", orient.z: " << odom.pose.pose.orientation.z
-                     << ", lin_est: " << odom.twist.twist.linear.x
-                     << ", ang_est: " << odom.twist.twist.angular.z);
-    last_odom = odom;
-  }
-};
-
-tf::Quaternion tfQuatFromGeomQuat(const geometry_msgs::Quaternion& quat)
-{
-  return tf::Quaternion(quat.x, quat.y, quat.z, quat.w);
-}
+#include "test_common.h"
 
 // TEST CASES
 TEST_F(DiffDriveControllerTest, testForward)
